+++ src/infra/db/mongodb/account/account-mongo-repository.ts
@@ -1,10 +1,12 @@
-import { AddAccountRepository } from '../../../../data/protocols/db/add-account-repository'
-import { LoadAccountByEmailRepository } from '../../../../data/protocols/db/load-account-by-email-repository'
+import { ObjectId } from 'mongodb'
+import { AddAccountRepository } from '../../../../data/protocols/db/account/add-account-repository'
+import { LoadAccountByEmailRepository } from '../../../../data/protocols/db/account/load-account-by-email-repository'
+import { UpdateAccessTokenRepository } from '../../../../data/protocols/db/account/update-access-token-repository'
 import { AccountModel } from '../../../../domain/models/account'
 import { AddAccountModel } from '../../../../domain/usecases/add-account'
 import { MongoHelper } from '../helpers/mongo-helper'
 
-export class AccountMongoRepository implements AddAccountRepository, LoadAccountByEmailRepository {
+export class AccountMongoRepository implements AddAccountRepository, LoadAccountByEmailRepository, UpdateAccessTokenRepository {
   async add (accountData: AddAccountModel): Promise<AccountModel> {
     const accountColletion = await MongoHelper.getCollection('accounts')
     const result = await accountColletion.insertOne(accountData)
@@ -17,5 +19,10 @@ export class AccountMongoRepository implements AddAccountRepository, LoadAccount
     const accountColletion = await MongoHelper.getCollection('accounts')
     const result = await accountColletion.findOne({ email })
     return result && MongoHelper.map(result)
+  }
+
+  async updateAccessToken (id: string, token: string): Promise<void> {
+    const accountColletion = await MongoHelper.getCollection('accounts')
+    await accountColletion.updateOne({ _id: new ObjectId(id) }, { $set: { accessToken: token } })
   }
 }
