+++ src/main/decorators/log-controller-decorator.spec.ts
@@ -1,7 +1,7 @@
 import { HttpRequest, HttpResponse, Controller } from '../../presentation/protocols'
 import { serverError } from '../../presentation/helpers/http/http-helper'
-import { LogControllerDecorator } from './log'
-import { LogErrorRepository } from '../../data/protocols/db/log-error-repository'
+import { LogControllerDecorator } from './log-controller-decorator'
+import { LogErrorRepository } from '../../data/protocols/db/log/log-error-repository'
 
 const makeLogErrorRepository = (): LogErrorRepository => {
   class LogErrorRepositoryStub implements LogErrorRepository {
